Move IsLens into an internal module.
