+++ fresnel/src/Fresnel/Lens.hs
@@ -15,25 +15,17 @@ module Fresnel.Lens
 , unpackedLens
 ) where
 
-import Control.Arrow (Kleisli, (&&&), (***))
+import Control.Arrow ((&&&), (***))
 import Data.Profunctor
 import Fresnel.Iso (IsIso)
+import Fresnel.Lens.Internal (IsLens)
 import Fresnel.Optic
-import Fresnel.Profunctor.OptionalStar (OptionalStar)
 
 -- Lenses
 
 type Lens s t a b = forall p . IsLens p => Optic p s t a b
 
 type Lens' s a = Lens s s a a
-
-class (IsIso p, Strong p) => IsLens p
-
-instance IsLens (->)
-instance Monad m => IsLens (Kleisli m)
-instance IsLens (Forget r)
-instance Functor f => IsLens (Star f)
-instance Functor f => IsLens (OptionalStar f)
 
 
 -- Construction
